requirements.txt
discord.py==2.6.0
pytz